der: increase `Length::MAX` to 256 MiB
There are legitimate cryptographic use cases that need to support
ASN.1 DER-encoded documents larger than the previous maximum of 1 MiB.

Two notable applications are where messages can be quite large are:

- Certificate Revocation Lists (CRLs)
- Cryptographic Message Syntax (CMS) a.k.a. PKCS#7

This increases the maximum length to something which is hopefully still
somewhat conservative but also useful enough for the above applications.

It can be further increased if there is a legitimate use case.