sec1: `Elliptic-Curve-Point-to-Octet-String` encoding
Adds support for encoding elliptic curve points according to the
following algorithms defined in SEC1:

- `Elliptic-Curve-Point-to-Octet-String`
- `Octet-String-to-Elliptic-Curve-Point`

These encodings use a tag byte (0x02/0x03 for compressed, 0x04 for
uncompressed) followed by one or two big endian-serialized field
elements representing the x/y coordinates of a curve point.

This is an extraction of the implementation from the `elliptic-curve`
crate, decoupled from the types in that crate and thus generally useful
for any crate which wishes to implement SEC1 support.