der: add `SequenceOf` type
Adds a type for `SequenceOf`, backed by the internal `ArrayVec` type.

This allows for array-backed storage of an ASN.1 SEQUENCE OF, where the
array size acts as a maximum, rather than an exact length requirement.

Support for modeling `[T; N]` as a SEQUENCE OF is retained, and now
implemented using `SequenceOf` to handle decoding.