der: remove `SetOf` trait; rename `SetOfArray` => `SetOf`
The `der` crate now has SET OF types which cover all cases:

- `SetOf(Array)`: heapless use case
- `BTreeSet`: heap-allocated use case

The `SetOf` trait was intended to provide a way to write generic code
that wasn't concerned with the actual storage of a decoded SET OF, with
the idea that a heapless crate user would impl it on their own data
structures.

Now that `der` provides a heapless data structure, users can simply pick
whether they want to use `SetOf` or use `BTreeSet`.