der: add `Document` trait
Adds a trait for representing ASN.1 DER-encoded documents, including
generalized support for PEM encoding as well as reading/writing such
documents from/to the underlying filesystem.

The intended usage is to DRY out the `*Document` types that exist in the
`pkcs1`, `pkcs8`, and `sec1` crates.

As a prototype/POC, this commit also impls the trait for `RsaPublicKey`
in the `pkcs1` crate. The rest are left as a followup exercise.