pkcs1: add support for multi-prime RSA keys
Implements the `OtherPrimeInfo` ASN.1 schema and with it full-fledged
support for multi-prime RSA keys.

Support is gated on the `alloc` feature as the `SEQUENCE OF` containing
`OtherPrimeInfos` is modeled as a `Vec`.

On heapless targets, a non-public dummy `OtherPrimeInfos` struct is used
instead.

Includes a test vector generated by the following version of OpenSSL:

    OpenSSL 1.1.1k  25 Mar 2021

Key was generated using the following command:

    $ openssl genrsa -primes 3 -out rsa2048-priv-3prime.pem 2048