spki: move `PublicKeyDocument` from `pkcs8` crate
This type provides heap-backed storage for `SubjectPublicKeyInfo`.

It was placed in the `pkcs8` crate instead of the `spki` crate because
that's where PEM support used to live.

Now that PEM support has been extracted into the `pem-rfc7468` crate,
this type can move to the `spki` crate.

To provide backwards compatibility, it's re-exported from the `pkcs8`
crate (where it's effectively the associated "public key type").