der: add support for hardened `Document` file permissions
When persisting `der::Document` as either DER or PEM, and the document
contains a secret (e.g. cryptographic key), this commit adds support for
using hardened file permissions on Unix platforms to prevent accidental
leakage to other users on the same host.

The `pkcs1`, `pkcs8`, and `sec1` crates already use such permissions,
however this change allows the file permission handling to be
implemented in a single place, so adding support for hardened
permissions on other OSes (Windows would be desirable) can happen in a
single location.

The aforementioned crates have not yet been updated to use this new API,
but it is at least now possible.