der: replace `ErrorKind::Truncated` with `ErrorKind::Incomplete` (#143)
Renames and replaces the previous `ErrorKind::Truncated` which carried
no additional information with an `ErrorKInd::Incomplete` which carries
the expected and actual lengths.

This is useful for incremental processing where a reader might attempt
to perform an I/O operation to further fill a buffer in the event that
the document is incomplete.

This commit includes a number of incidental changes to handle cases
where `ErrorKind::Truncated` was inappropriately used before. One of the
biggest is adding an error type to the `spki` crate in order to better
signal the error conditions (e.g. `AlgorithmParametersMissing`) which
were previously unhelpfully propagated as `der::ErrorKind::Truncated`.