der_derive: don't automatically derive `From` impls for `Choice` (#168)
Previously the custom derive for `Choice` would also write `From` impls
for each of the variants, which would convert the inner type into a
particular variant of the enum. This used to be an integral part of the
`Decode` impl, but is no longer.

This is problematic because it means no two variants can have the same
inner type. This commit gets rid of these derives since they're no
longer needed for decoding and have this limitation.