der: add test for `SetOf` regression
These were failing to parse correctly until inadvertently fixed by #178.

The problem is not fully addressed, but it's good to at least capture it
in a regression test to prevent future breakages.

Note that this test also seems to be uncovering another bug in
`const-oid` which will get separately addressed in a followup.