der: rename `Tagged` => `FixedTag`; add new `Tagged` trait
This commit does the following:

- Renames the old `Tagged` trait to `FixedTag`, to reflect that the tag
  for that type is constant regardless of its value.
- Introduces a new `Tagged` trait with a `fn tag(&self) -> Tag` method.
  This allows types like `Any` and `ContextSpecific` to be able to
  participate in protocols that need to know a given type's tag.
- Bounds `Choice` on `Tagged`.

`Tagged` has a blanket impl for all `FixedTag` types.

This change is pretty necessary for determining DER ordering for
`SET OF`, as it allows for constructing the `Header` of any type which
is `EncodeValue + Tagged`, which is needed to compare `Header`s.

The `Header` is also useful for generalizing `Encodable`, and indeed
this commit removes the handwritten impl of `Encodable` for both `Any`
and `ContextSpecific` since the blanket impls now write them
automatically.