der_derive: remove `synstructure` dependency
Little of `synstructure`'s actual functionality is being used, and given
all of the additional parsing work that needs to be done in the proc
macro, we already have the relevant context that `synstructure` would
abstract over.

Removing `synstructure` also means we can start using the
`proc-macro-error` crate to provide better error messages.