pem-rfc7468 v0.2.4
This commit adds a changelog entry about `pem-rfc7468` v0.2.4, which was
released to address RustCrypto/RSA#116.

The release is identical to v0.2.3, but with `base64ct` restricted to
`<1.2` in order to prevent MSRV-related breakages.

Since `master` is already on `pem-rfc7468` v0.3.0-pre which is also a
2021 edition crate, those changes are not actually included in this
commit, but rather just the changelog entry.