der_derive: add "IR" to all derivation types
Adds an "intermediate representation" for all custom derive macros which
uses plain Rust structs to store the relevant parameters to the custom
derive.

This makes it much easier to write custom derive unit tests, since the
tests can assert on the IR. This commit also includes unit tests for all
of the examples in `der/tests/derive.rs` to ensure that the IR is in the
expected state.

The lowerings to `TokenStream` are moved to the IR types.