der: add `ContextSpecificRef` type
Adds a reference type for encoding context-specific fields, and unifies
all context-specific encoding under that.

This also makes it possible to construct a context-specific encoder
trait object for use with `Sequence::fields`.

Additionally, this change fixes a TODO to check that the constructed bit
is set correctly for IMPLICIT context-specific fields.