der: use `DerOrd` for `SetOf*` types
Changes `SetOf` to use the `DerOrd` trait instead of `Ord`.

Removes support for using `BTreeSet` as a `SET OF`, since we need to
respect `DerOrd` instead of `Ord`.

Replaces it with a heap-backed `SetOfVec` type that works like `SetOf`
but backed by a `Vec` and therefore with support for an arbitrary number
of elements.