ssh-key: add test vectors to `tests/examples` (#260)
Adds a set of private and public keys for each supported algorithm
generated by the `ssh-keygen` tool from OpenSSH_8.1p1

A `generate.sh` script which was used to drive the generation process is
also included.

The key fingerprints for the generated keys are as follows:

- id_dsa_1024.pub: SHA256:Nh0Me49Zh9fDw/VYUfq43IJmI1T+XrjiYONPND8GzaM
- id_ecdsa_p256.pub: SHA256:JQ6FV0rf7qqJHZqIj4zNH8eV0oB8KLKh9Pph3FTD98g
- id_ecdsa_p384.pub: SHA256:nkGE8oV7pHvOiPKHtQRs67WUPiVLRxbNu//gV/k4Vjw
- id_ecdsa_p521.pub: SHA256:l3AUUMK6Q2BbuiqvMx2fs97f8LUYq7sWCAx7q5m3S6M
- id_ed25519.pub: SHA256:UCUiLr7Pjs9wFFJMDByLgc3NrtdU344OgUM45wZPcIQ
- id_rsa_3072.pub: SHA256:Fmxts/GcV77PakFnf1Ueki5mpU4ZjUQWGRjZGAo3n/I
- id_rsa_4096.pub: SHA256:FKAyeywtQNZLl1YTzIzCV/ThadBlnWMaD7jHQYDseEY

No passphrase was supplied. An additional set of test vectors will need
to be generated for testing password-based encryption.