base64ct: add stateful `Decoder` type
Adds a buffered `Decoder` which is able to incrementally decode an
output stream from an unpadded input.

Presently constrained to contiguous Base64 inputs which do not contain
newlines, although an important next step is to encapsulate line-wrapped
inputs.

Support for padded inputs is untested and broken, but will be fixed in a
followup commit.