base64ct: add `Decoder::new_linewrapped`
Adds a mode to the `Decoder` type for decoding Base64 which is wrapped
at a specific width.

The internal `BlockBuffer` is used to handle Base64 input blocks which
are spread across multiple lines, so any line width greater than 0 is
supported. This is important because the OpenSSH private key format
linewraps at a width of 70, which is not a multiple of 4.