base64ct: fix `Decoder` handling of multi-line padded Base64
This fix uses a similar approach of eagerly unpadding the input Base64
used when handling contiguous unwrapped Base64.

This is probably a naive strategy though: it's possible that the padding
might end up getting split over a newline boundary.

A better approach would be to use a padded decoder when processing the
last block of input Base64. That would require a bit more work to ensure
that the correct Base64 variant is being used given the context.