spki: add `Error::KeyMalformed` variant
Adds an error variant similar to `pkcs8::Error::KeyMalformed` for
relaying that binary data contained in `subject_public_key` is in some
way malformed beyond what errors can be communicated via `Error::Asn1`
for DER decoding errors.

This is useful if DER decoding succeeds but something else is wrong,
e.g. an elliptic curve public key is not a solution to the curve
equation.