Proposed support of REAL DER type
Covers sections 8.5 and 11.3 of ITU-T X.690 (02/2021)

Caveats:
	1. Interpretation of section 11.3.1 to be further discussed
	2. Currently, all encoding is binary (no ISO 6093 NR3)

Testing of arbitrarily large and small f64s shows that DER typically
requires 12 bytes to encode these 8-byte IEEE-754 values. This can be
explained because of the extra two bytes for the tag and length, and the
fact that the exponent is encoded on 1 to 2 bytes (only 11 bits in
IEEE-754), and the mantissa may require up to 8 bytes itself, hence 12
bytes total. Of note that, depending on Caveat 1 above, this information
may be wrong.