der_derive: extract `LowerFieldDecoder`
Extracts an internal stateful `TokenStream` newtype which handles
transforming the AST-in-progress.

In somewhat handwavy MVC terms, if `DeriveSequence` is the "model", and
`SequenceField` the controller, then `LowerFieldDecoder` is the view.

This should eventually help with more composable orderings that can
better combine different features, however for now it retains many of
the shortcomings (and likewise compatibility with) the previous
implementation.