der: make `OrdIsValueOrd` into sealed trait
While this trait is useful for defining the `ValueOrd` impl for various
ASN.1 universal types within the `der` crate itself, it is easily abused
outside the context of the `der` crate.

`ValueOrd` handles comparing DER serializations for ordering. The
`OrdIsValueOrd` is tempting in that it makes it appear that `Ord` can be
used instead, but this only works in hyper-specific cases where `Ord` is
identical to `ValueOrd`, which is primarily the ASN.1 universal types.

To prevent misuse, this commit seals the trait to force downstream users
to impl `ValueOrd`.