der: add `SequenceRef` type
Adds a type which wraps the value portion / "body" of a DER-serialized
`SEQUENCE`.

This type subsumes the previous decoding logic for `Decoder::sequence`,
parsing the `SEQUENCE` using `DecodeValue` and storing its position
within the outer document.

In addition to simplifying the sequence decoding logic, this also makes
it possible to use this type within the `Sequence` proc macro, which
will allow it to impl `DecodeValue`. However, that work is left for a
followup PR.