der_derive: have `Sequence` macro derive `DecodeValue`
The proc macro previously derived the `Decode` trait, which is not
composable with context-specific fields, which depend on the
`DecodeValue` trait in order to support IMPLICIT fields.

This commit changes the custom derive macro for `Sequence` to produce
a `DecodeValue` impl instead of `Decode`. This leverages the newly
introduced `SequenceRef` type from #374.