der: impl `From`/`FromIterator` for `SetOfVec`
Adds impls of these traits for `SetOfVec`, which sort elements using
`DerOrd`. The `FromIterator` impl is composed in terms of the
`From<Vec<T>>` impl, which collects the elements into a `Vec`, then
sorts them with `sort_by`.

One problem is `DerOrd::der_cmp` is fallible, so in order to use it with
`sort_by` the only option is to panic if it returns an `Err`. I'm not
sure if this can be resolved short of shipping a custom sort function or
getting something like `try_sort_by` added to `core`. FWIW I proposed
the latter here:

https://internals.rust-lang.org/t/fallible-sorting/16066