der: support `SetOf` sorting on heapless `no_std` targets
Rust's built-in `[T]::sort_by` requires both `alloc` and a panic
handler, which is limiting for `no_std` targets.

This commit switches to using a heapless-friendly insertion sort
implementation which is also fallible and can bubble up errors from
`DerOrd::der_cmp`.

Insertion sort is small (<10loc) and should perform well when the input
is mostly in order, which should hopefully often be the case for `SetOf`
constructors. It also works completely in-place and is thus friendly to
heapless targets.

This PR replaces the previous `[T]::sort_by` approach used by
`SetOfVec`, which included a workaround which handled `DerOrd::der_cmp`
errors by returning `Ordering::Less` and doing a second pass over the
results to check they were correctly ordered. This might provide better
performance and be worth bringing back, but for now this PR chooses
instead to consolidate on a single, well-tested implementation.

Testing is performed using proptest, to check equivalence with
Rust's built-in `[T]::sort_by`.