base64ct: add bounds for `Encoding`/`Variant` ZSTs
Deriving traits like `Copy`, `Clone`, and `Debug` on types which include
an encoding as a generic parameter is presently problematic because
there are no trait bounds specified.

These traits, `Encoding` and the sealed `Variant` which receives a
blanket impl of the former, are not intended to be implemented by types
outside of this crate. While adding bounds is technically a breaking
change, it would be highly unexpected for this to have any real-world
impact as external crates should only be consuming these traits and
using them with the encoding ZSTs defined in this crate.