der: fix handling of oversized unsigned `INTEGER` inputs (#447)
The previous implementation used `saturating_sub` rather than
`checked_sub` to compute the number of leading zeroes to use, which
would cause a panic if the input exceeded the output (see #446).

This commit switches to `checked_sub`, returning `ErrorKind::Length` in
the event the output buffer is too small for the given input. It also
adds unit tests for this behavior as well as the happy paths.