base64ct: rename `Decoder::decoded_len` => `::remaining_len`
This is technically a breaking change, but this method was only just
introduced in the v1.4.0 release which occurred less than 24h ago, so
changing things now, yanking that release, and cutting a replacement
v1.4.1 release should hopefully not cause any real-world breakages.

In using this method it felt confusingly named: it returns the number of
bytes of as-yet-to-be-decoded data, but the "decoded" name implies past
tense, which feels more like it's returning the number of bytes which
have been decoded so far.

It *would* be useful to have such a method as well which tracks the
current cursor/position within the decoded output, but adding one would
be rather confusing as `decoded_len` would be a good candidate for
*that* name.