ssh-key: make `authorized_keys` types owned
Owned types are easier to work with.

This commit retains the config options parser in a heapless mode but
requires `alloc` to actually store them.

Since they were the only borrowed part of `authorized_key::Entry`, it
can become owned as well.

This allows `AuthorizedKeys::read_file` to return a `Vec<Entry>` rather
than taking a callback that borrows a string.