ssh-key: consolidate `KdfAlg` and `KdfOpts` into `Kdf`
Since the KDF algorithm and options are located side-by-side in private
keys, and this crate models them as effectively a single `enum`, this
commit makes that a first-class arrangement, combining the two into a
unified `pub enum Kdf`.

This allows the enum to impl the `Decode` trait proper.

Internally it still decodes/encodes a `KdfAlg`, however it only uses it
to select the enum variant and inform the KDF options decoding.