ssh-key: add `Decoder::decode_length_prefixed`
Adds a helper method for decoding length-prefixed data which calls a
provided decoder function/closure and then afterward validates that the
correct amount of data was read.

This could potentially be adapted to use a special decoder wrapper which
enforces that data can't be read past the end of the buffer, but for now
it uses a late check after the inner decoder function is finished.