pem-rfc7468: use `str::from_utf8_unchecked` in `encode`
There's a unsafety versus sidechannel resistence tradeoff in converting
PEM output by the `Encoder` type to a `str` (which performs branch-heavy
UTF-8 validity checks).

This commit introduces a tiny bit of `unsafe` code (the very first
`unsafe` code in this crate) to bypass these checks.

In their place, there is both a `debug_assert!` which performs the full
`str::from_utf8` check when `debug_assertions` are enabled, as well as
an always-on constant time-ish runtime check to ensure characters in the
output document are all in the 7-bit ASCII set.