der: `Document` and `SecretDocument` types
Replaces the previous `Document` trait with types for wrapping
serialized ASN.1 DER SEQUENCEs stored on the heap.

The `SecretDocument` is composed in terms of `Document` and provides
some additional hardening for sensitive data, namely zeroize-on-drop and
file permissions hardening when writing to disk.

This commit also removes all of the format-specific `*Document` types
found in the `pkcs1`, `pkcs8`, `spki`, and `sec1` crates with the new
`Document` and `SecretDocument` types.