ssh-key: preliminary certificate decoder support
Initial decode-only support for OpenSSH certificates. This is
preliminary work which does not expose much of the certificate data as
part of the public API, as types needed to properly model the
certificate still need to be designed/implemented.

Includes example SSH certificates generated from the already checked in
example public keys, generated with the following commands:

$ ssh-keygen -s id_ed25519 -I ed25519 id_ed25519.pub
$ ssh-keygen -s id_ed25519 -I ecdsa-nistp256 -h -Z host.example.com -V -2d:+28452d id_ecdsa_p256.pub
$ ssh-keygen -s id_ed25519 -I dsa -V -2d:+28452d id_dsa.pub
$ ssh-keygen -s id_ed25519 -I dsa -V -2d:+28452d id_dsa_1024.pub
$ ssh-keygen -s id_ed25519 -I rsa -V -2d:+28452d id_rsa_4096.pub