ssh-key: private key checkint improvements (#577)
This commit adds support for storing the "checkint" for a private key
in the `PrivateKey` type. This makes it possible to reserialize a
decoded private key byte-for-byte.

Additionally, encryption and random key generation now generate a random
"checkint".

The fallback deterministic method is moved from the public to private
key, ensuring that a "checkint" is always secret. This should help
prevent hypothetical bitflipping attacks which would be possible if the
deterministic checkint were computed from public data.