der: add `EncodeRef`/`EncodeValueRef`; remove `OptionalRef`
Adds generic wrapper types for encoding references to types which impl
encoding traits:

- `EncodeRef`: impls `Encode` for &T where T: `Encode`
- `EncodeValueRef`: impls `EncodeValue` and `Tagged` for &T

This, in conjunction with impls for `&Option<T>`, eliminates the need
to have a dedicated `OptionalRef` type.

The `ContextSpecificRef` type is retained as it's a useful constructor
for use in proc macros.

All in all it would be nice to have a better solution like this, like a
blanket impl of `Encode` for &T where T: `Encode`. That's a direction
which may still be worth exploring, but in the meantime the wrapper
types mean there's a generic solution for reference type encoding.