der: make `Reader::remaining_len` infallible
Like #607, this makes a length calculation infallible, and also adds a
default implementation.

It uses saturating arithmetic to avoid panics, with `debug_assert!` to
make sure the operands of the calculation make sense.

That said, if there is a bug where a reader's cursor/position exceeds
the `input_len()`, it will still calculate a value that makes sense for
`remaining_len()` when using saturating arithmetic: 0.