der: use `&mut dyn Writer` as output for `Encode::encode`
Implements encoding in terms of the `Writer` trait.

To allow `Encode` to continue to function as a trait object, the
`Writer` must also be a trait object.

This approach should enable 1-pass PEM encoding as well as computing key
fingerprints without an output buffer. It might come at a cost to
performance due to the increased use of trait objects.

It might be possible to retain monomorphization using
`Encode<W: Writer>` but at a cost to ergonomics. However, that might be
worth exploring depending on the performance.