spki: streaming fingerprint builder (#616)
Uses the `Writer` trait to compute SPKI fingerprints by hashing output
DER on-the-fly as opposed to encoding it first to an intermediate buffer
then computing the SHA-256 digest.

Besides simply being more efficient, the buffered method had a hardcoded
limit of 4096-bytes. The new implementation has no such limit because it
hashes data as soon as it's encoded.

This additionally renames `SubjectPublicKeyInfo::fingerprint` to
`::fingerprint_bytes`. "True" SPKI fingerprints are Base64-encoded, and
this commit also promotes `fingerprint_base64` to being earlier in the
impl block so it's the first fingerprinting method people see.