pem-rfc7468: add `encapsulated_len_wrapped`
Estimates upper bound of the length of encoded line-wrapped Base64 with
a specified customizable line width and line ending.

Also changes the input parameters of `encapsulated_len`, which now
accepts the length of the raw input data prior to Base64 encoding,
making it significantly less error prone.

Together these functions are used to DRY out similar calculations
presently occurring in the `der` and `ssh-key` crates.

This is a breaking change, so it also bumps the version to v0.6.0-pre.