Allow empty comments in openssh keys.
Even the comments say this is allowed. This also, however means
that it's not possible to detect a truncated key just from the end
of input, so that case is now treated as the end of the key.

Detection of truncated keys should probably happen at a higher
level, where the correct length of the key is known.