der: remove decode-time SET OF ordering checks
Some DER serializer implementations fail to properly sort elements of a
`SET OF`. This is technically non-canonical, but occurs frequently
enough that most DER decoders tolerate it. Unfortunately because
of that, we must also follow suit.

This commit builds an `ArrayVec` or `Vec` of the incoming decoded set
elements, and then unconditionally sorts it retroactively, ensuring all
`SET OF` types internally reflect the proper order, but out-of-order
`SET OF` messages are tolerated when coming off-the-wire.

The sort implementation is merge sort, which should be efficient for
already-in-order or mostly-in-order sets.