der+x509: replace `Newtype` custom derive macro with macro_rules! (#626)
In attempting to implement the `Decode`/`DecodeValue` changes for #525,
I ran into trouble trying to modify the `Newtype` derive macro. In its
present form it both does quite a bit which makes it difficult to test
the lowerings. As the lowerings aren't well-tested, it becomes extremely
difficult to debug when the proc macro is emitting bad tokens.

In looking at those problems, I came to the conclusion that a
`macro_rules!` macro would be better for several reasons:

- Easier to modify and debug
- There is no corresponding `Newtype` trait, which breaks with the
  spirit of custom derive macros
- The `Newtype` macro was stealthily defining several macro as the same
  time, which makes it somewhat inscrutible (especially with the absence
  of a same-named trait which bounds on the traits it impls).
- At least within the context of this repo, this is a pattern specific
  to the `x509` crate.

If we want to go the custom derive route, it would be better to break
it down into separate trait-specific macros for e.g.
`DecodeValue`/`EncodeValue` which are easier to test/debug and better
map to the custom derive pattern.