pkcs1: cleanups to `params` module
Followup to #698.

- Renames `RsaPSSParameters` => `RsaPssParams`
- Removes use of the `der` crate's `derive` feature

The `rsa` crate curently doesn't need any custom derive, and adding it
here would make it a hard requirement.

We can consider switching to custom derive in `pkcs1`, but if we do it
then it should be wholesale for all types in the crate, not using custom
derive for some types and handwritten impls for others.