const-oid: generic backing buffer (#757)
The current buffering approach, with a fixed-sized `ArrayVec`-like
backing buffer, is helpful for being able to use `const fn` to be able
to parse OID strings at compile time and re-encode them in BER form.
This functionality is preserved as a new `Buffer` type, which still
supports all current `const fn` functionality.

`ObjectIdentifier` has been changed to be generic around
`B: AsRef<[u8]>` which allows a `const fn` constructor that accepts a
BER-encoded slice.

This is useful for when there are large numbers of OID constants, as in
the `const_oid::db` module. However, while this module has been updated
so the code compiles, it has not yet been changed to take advantage of
this new approach.