[WIP] der: eliminate dynamism from encoding
Replaces use of `dyn Writer` with `impl Writer`.

Removes the `Sequence::fields` method, which was used for a blanket
impl of `EncodeValue`.

Instead, very small changes to the custom derive can be used to derive a
monomorphic implementation of `EncodeValue` that doesn't rely on trait
objects, where calls to the `Writer` can be potentially inlined.